+++ spec/chewy/index/actions_spec.rb
@@ -15,8 +15,8 @@ describe Chewy::Index::Actions do
   end
 
   describe '.create' do
-    specify { expect(DummiesIndex.create["acknowledged"]).to eq(true) }
-    specify { expect(DummiesIndex.create('2013')["acknowledged"]).to eq(true) }
+    specify { expect(DummiesIndex.create['acknowledged']).to eq(true) }
+    specify { expect(DummiesIndex.create('2013')['acknowledged']).to eq(true) }
 
     context do
       before { DummiesIndex.create }
@@ -31,7 +31,7 @@ describe Chewy::Index::Actions do
       specify { expect(DummiesIndex.aliases).to eq([]) }
       specify { expect(DummiesIndex.indexes).to eq(['dummies_2013']) }
       specify { expect(DummiesIndex.create('2013')).to eq(false) }
-      specify { expect(DummiesIndex.create('2014')["acknowledged"]).to eq(true) }
+      specify { expect(DummiesIndex.create('2014')['acknowledged']).to eq(true) }
 
       context do
         before { DummiesIndex.create '2014' }
@@ -49,8 +49,8 @@ describe Chewy::Index::Actions do
   end
 
   describe '.create!' do
-    specify { expect(DummiesIndex.create!["acknowledged"]).to eq(true) }
-    specify { expect(DummiesIndex.create!('2013')["acknowledged"]).to eq(true) }
+    specify { expect(DummiesIndex.create!['acknowledged']).to eq(true) }
+    specify { expect(DummiesIndex.create!('2013')['acknowledged']).to eq(true) }
 
     context do
       before { DummiesIndex.create }
@@ -79,7 +79,7 @@ describe Chewy::Index::Actions do
         skip_on_version_lt('2.0', 'format of exception was changed')
         expect { DummiesIndex.create!('2013') }.to raise_error(Elasticsearch::Transport::Transport::Errors::BadRequest).with_message(/index_already_exists_exception.*dummies_2013/)
       end
-      specify { expect(DummiesIndex.create!('2014')["acknowledged"]).to eq(true) }
+      specify { expect(DummiesIndex.create!('2014')['acknowledged']).to eq(true) }
 
       context do
         before { DummiesIndex.create! '2014' }
@@ -102,7 +102,7 @@ describe Chewy::Index::Actions do
 
     context do
       before { DummiesIndex.create }
-      specify { expect(DummiesIndex.delete["acknowledged"]).to eq(true) }
+      specify { expect(DummiesIndex.delete['acknowledged']).to eq(true) }
 
       context do
         before { DummiesIndex.delete }
@@ -112,7 +112,7 @@ describe Chewy::Index::Actions do
 
     context do
       before { DummiesIndex.create '2013' }
-      specify { expect(DummiesIndex.delete('2013')["acknowledged"]).to eq(true) }
+      specify { expect(DummiesIndex.delete('2013')['acknowledged']).to eq(true) }
 
       context do
         before { DummiesIndex.delete('2013') }
@@ -122,7 +122,7 @@ describe Chewy::Index::Actions do
 
       context do
         before { DummiesIndex.create '2014' }
-        specify { expect(DummiesIndex.delete["acknowledged"]).to eq(true) }
+        specify { expect(DummiesIndex.delete['acknowledged']).to eq(true) }
 
         context do
           before { DummiesIndex.delete }
@@ -147,7 +147,7 @@ describe Chewy::Index::Actions do
 
     context do
       before { DummiesIndex.create }
-      specify { expect(DummiesIndex.delete!["acknowledged"]).to eq(true) }
+      specify { expect(DummiesIndex.delete!['acknowledged']).to eq(true) }
 
       context do
         before { DummiesIndex.delete! }
@@ -157,7 +157,7 @@ describe Chewy::Index::Actions do
 
     context do
       before { DummiesIndex.create '2013' }
-      specify { expect(DummiesIndex.delete!('2013')["acknowledged"]).to eq(true) }
+      specify { expect(DummiesIndex.delete!('2013')['acknowledged']).to eq(true) }
 
       context do
         before { DummiesIndex.delete!('2013') }
@@ -167,7 +167,7 @@ describe Chewy::Index::Actions do
 
       context do
         before { DummiesIndex.create '2014' }
-        specify { expect(DummiesIndex.delete!["acknowledged"]).to eq(true) }
+        specify { expect(DummiesIndex.delete!['acknowledged']).to eq(true) }
 
         context do
           before { DummiesIndex.delete! }
@@ -187,8 +187,8 @@ describe Chewy::Index::Actions do
   end
 
   describe '.purge' do
-    specify { expect(DummiesIndex.purge["acknowledged"]).to eq(true) }
-    specify { expect(DummiesIndex.purge('2013')["acknowledged"]).to eq(true) }
+    specify { expect(DummiesIndex.purge['acknowledged']).to eq(true) }
+    specify { expect(DummiesIndex.purge('2013')['acknowledged']).to eq(true) }
 
     context do
       before { DummiesIndex.purge }
@@ -234,8 +234,8 @@ describe Chewy::Index::Actions do
   end
 
   describe '.purge!' do
-    specify { expect(DummiesIndex.purge!["acknowledged"]).to eq(true) }
-    specify { expect(DummiesIndex.purge!('2013')["acknowledged"]).to eq(true) }
+    specify { expect(DummiesIndex.purge!['acknowledged']).to eq(true) }
+    specify { expect(DummiesIndex.purge!('2013')['acknowledged']).to eq(true) }
 
     context do
       before { DummiesIndex.purge! }