+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Changes
 
+  * Brand new import `:bulk_size` option, set desired ElasticSearch bulk size in bytes.
+
   * Configurable per-type default import options (@barthez, #341)
 
   * Various codebase optimizations (@DNNX, @pyromaniac)